refactor : s3 설정 디렉토리 생성
